Flask test  and updated pipelines

+++ scripts/pipeline_region_start_end_combined.sh
@@ -200,7 +200,18 @@ if [[ $? -ne 0 ]]; then
     exit 1
 fi
 
-cp -P $tmp_dir/baseline/* $output_dir
+docker run $DOCKERGPUPARAMS --rm -it -m 32000m -v $input_dir:$input_dir -v $output_dir:$output_dir docker.laypa:latest \
+    python run.py \
+    -c configs/segmentation/region/region_dataset_imagenet_freeze.yaml \
+    -i $input_dir \
+    -o $output_dir \
+    --opts MODEL.WEIGHTS "" TEST.WEIGHTS pretrained_models/region_model_best_mIoU.pth
+    # > /dev/null
+
+if [[ $? -ne 0 ]]; then
+    echo "Region detection has errored, stopping program"
+    exit 1
+fi
 
 docker run $DOCKERGPUPARAMS --rm -it -m 32000m -v $output_dir:$output_dir -v $tmp_dir:$tmp_dir docker.laypa:latest \
     python utils/combine_start_end.py \
@@ -217,9 +228,8 @@ fi
 
 docker run --rm -v $output_dir:$output_dir -v $tmp_dir:$tmp_dir docker.loghi-tooling /src/loghi-tooling/minions/target/appassembler/bin/MinionExtractBaselinesStartEndNew3 \
     -input_path_png $output_dir/page/ \
-    -input_path_pagexml $tmp_dir/baseline/page/ \
-    -output_path_pagexml $output_dir/page/ \
-	-as_single_region true
+    -input_path_pagexml $output_dir/page/ \
+    -output_path_pagexml $output_dir/page/
 
 if [[ $? -ne 0 ]]; then
     echo "Extract baselines has errored, stopping program"
